src/agents/research_agent.py
"""
Web Research Agent for Multi-Skill Super-Agent

This module implements the Web Research Agent that searches the web,
scrapes content, and summarizes information based on user queries.
"""
import requests
from bs4 import BeautifulSoup
import openai
from loguru import logger

from .base_agent import Agent
from ..utils.config import OPENAI_API_KEY

class WebResearchAgent(Agent):
    """Agent for web research and summarization"""
    
    def __init__(self):
        """Initialize the web research agent"""
        super().__init__("WebResearch")
        openai.api_key = OPENAI_API_KEY
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
        }
    
    async def process(self, query: str) -> str:
        """
        Process a research query and return summarized information
        
        Args:
            query: The research query
            
        Returns:
            The summarized research results
        """
        try:
            # Pre-process the query
            processed_query = await self._pre_process(query)
            
            # Perform web search (simplified implementation)
            search_results = await self._search_web(processed_query)
            
            # Scrape content from top results
            content = await self._scrape_content(search_results[:3])
            
            # Summarize the content
            summary = await self._summarize_content(content, processed_query)
            
            # Post-process the response
            return await self._post_process(summary)
        
        except Exception as e:
            logger.error(f"Error in web research: {str(e)}")
            return f"Error performing research: {str(e)}"
    
    async def _search_web(self, query: str) -> list:
        """
        Perform a web search for the query
        
        Args:
            query: The search query
            
        Returns:
            A list of search result URLs
        """
        # This is a simplified implementation
        # In a real implementation, this would use a search API or web scraping
        logger.info(f"Searching web for: {query}")
        
        # Placeholder for search results
        # In a real implementation, these would be actual search results
        search_results = [
            "https://example.com/result1",
            "https://example.com/result2",
            "https://example.com/result3"
        ]
        
        return search_results
    
    async def _scrape_content(self, urls: list) -> str:
        """
        Scrape content from a list of URLs
        
        Args:
            urls: The list of URLs to scrape
            
        Returns:
            The combined content from all URLs
        """
        # This is a simplified implementation
        # In a real implementation, this would actually scrape the content
        logger.info(f"Scraping content from {len(urls)} URLs")
        
        combined_content = ""
        
        # In a real implementation, this would loop through the URLs and scrape content
        for url in urls:
            try:
                # Placeholder for scraped content
                # In a real implementation, this would be the actual scraped content
                content = f"Placeholder content from {url}\n"
                combined_content += content
            except Exception as e:
                logger.error(f"Error scraping {url}: {str(e)}")
        
        return combined_content
    
    async def _summarize_content(self, content: str, query: str) -> str:
        """
        Summarize content based on the original query
        
        Args:
            content: The content to summarize
            query: The original query
            
        Returns:
            The summarized content
        """
        try:
            # Create a system message for summarization
            system_message = (
                "You are a research assistant. "
                "Summarize the provided content to answer the user's query. "
                "Be concise but comprehensive, focusing on the most relevant information. "
                "Include key facts and cite sources when possible."
            )
            
            # Call OpenAI API to summarize content
            response = await openai.chat.completions.create(
                model="gpt-4",
                messages=[
                    {"role": "system", "content": system_message},
                    {"role": "user", "content": f"Query: {query}\n\nContent to summarize: {content}"}
                ],
                temperature=0.3,
                max_tokens=1000
            )
            
            # Extract the summary
            summary = response.choices[0].message.content
            
            return summary
        
        except Exception as e:
            logger.error(f"Error in summarization: {str(e)}")
            return f"Error summarizing content: {str(e)}"
