requirements.txt
# Python Telegram Bot
python-telegram-bot==20.7
# Web Framework
fastapi==0.109.2
uvicorn==0.27.1
# Agent Framework
langchain==0.1.9
langchain-openai==0.0.5
# Image Generation
openai==1.12.0
stability-sdk==0.8.5
# Web Research
beautifulsoup4==4.12.2
selenium==4.16.0
requests==2.31.0
# Task Automation
apscheduler==3.10.4
# Storage
redis==5.0.1
sqlalchemy==2.0.27
# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
loguru==0.7.2
# Testing
pytest==7.4.3
